# doc/swagger.yaml
openapi: 3.1.0
info:
  title: Readmaster.ai API
  description: API for Readmaster.ai, a web-based reading assessment platform.
  version: 0.1.0
paths:
...
  /api/v1/teacher/classes:
    post:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Create Class
      operationId: teacher_create_class_api_v1_teacher_classes_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassCreateDTO'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher List My Classes
      operationId: teacher_list_my_classes_api_v1_teacher_classes_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ClassResponseDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teacher/classes/{class_id}:
    get:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Get Class Details
      operationId: teacher_get_class_details_api_v1_teacher_classes__class_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Update Class Details
      operationId: teacher_update_class_details_api_v1_teacher_classes__class_id__put
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassUpdateDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Delete Class
      operationId: teacher_delete_class_api_v1_teacher_classes__class_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teacher/classes/{class_id}/students:
    post:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Add Student To Class
      operationId: teacher_add_student_to_class_api_v1_teacher_classes__class_id__students_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStudentToClassRequestDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher List Students In Class
      operationId: teacher_list_students_in_class_api_v1_teacher_classes__class_id__students_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDTO'
                title: Response Teacher List Students In Class Api V1 Teacher Classes  Class Id  Students Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teacher/classes/{class_id}/students/{student_id}:
    delete:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Remove Student From Class
      operationId: teacher_remove_student_from_class_api_v1_teacher_classes__class_id__students__student_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Student Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teacher/assignments/readings:
    post:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Assign Reading To Students
      operationId: teacher_assign_reading_to_students_api_v1_teacher_assignments_readings_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignReadingRequestDTO'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
  /api/v1/teacher/classes/{class_id}/progress-report:
    get:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Get Class Progress Report
      operationId: teacher_get_class_progress_report_api_v1_teacher_classes__class_id__progress_report_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassProgressReportDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teacher/students/{student_id}/progress-summary:
    get:
      tags:
        - Teacher - Class, Assignment & Progress
      summary: Teacher Get Student Progress Summary
      operationId: teacher_get_student_progress_summary_api_v1_teacher_students__student_id__progress_summary_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Student Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProgressSummaryDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teacher/students:
    post:
      tags:
        - Teacher - Account Management
      summary: Teacher Create Student Account
      description: Allows an authenticated teacher to create a new student account.
      operationId: teacher_create_student_account_api_v1_teacher_students_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherStudentCreateRequest'
      responses:
        '201':
          description: Student account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized. Invalid or missing authentication token.
        '403':
          description: Forbidden. Authenticated user is not a teacher.
        '422':
          description: Validation Error. Invalid input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AddStudentToClassRequestDTO:
      properties:
        student_id:
          type: string
          format: uuid
          title: Student Id
          description: The ID of the student to be added to the class.
      type: object
      required:
        - student_id
      title: AddStudentToClassRequestDTO
      description: DTO for adding a student to a class.
...
    AssignReadingRequestDTO:
      properties:
        reading_id:
          type: string
          format: uuid
          title: Reading Id
          description: The ID of the reading material to assign.
        student_ids:
          anyOf:
            - items:
                type: string
                format: uuid
              type: array
            - type: 'null'
          title: Student Ids
          description: A list of specific student IDs to assign the reading to.
        class_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Class Id
          description: The ID of a class to assign the reading to all its students. If provided, student_ids might be ignored or augmented.
        due_date:
          anyOf:
            - type: string
              format: date
            - type: 'null'
          title: Due Date
          description: Optional due date for the assignment.
      type: object
      required:
        - reading_id
      title: AssignReadingRequestDTO
      description: DTO for a teacher to assign a reading to students or a class.
...
    ClassCreateDTO:
      properties:
        class_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Class Name
          description: Name of the class.
        grade_level:
          anyOf:
            - type: string
              maxLength: 50
            - type: 'null'
          title: Grade Level
          description: Grade level of the class, e.g., 'Grade 5', 'Secondary II'.
      type: object
      required:
        - class_name
      title: ClassCreateDTO
      description: |-
        DTO for creating a new class.
        'created_by_teacher_id' will be derived from the authenticated teacher user in the use case.
...
    ClassResponseDTO:
      properties:
        class_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Class Name
          description: Name of the class.
        grade_level:
          anyOf:
            - type: string
              maxLength: 50
            - type: 'null'
          title: Grade Level
          description: Grade level of the class, e.g., 'Grade 5', 'Secondary II'.
        class_id:
          type: string
          format: uuid
          title: Class Id
        created_by_teacher_id:
          type: string
          format: uuid
          title: Created By Teacher Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        students:
          items:
            $ref: '#/components/schemas/UserResponseDTO'
          type: array
          title: Students
          description: List of students enrolled in the class.
          default: []
      type: object
      required:
        - class_name
        - class_id
        - created_by_teacher_id
        - created_at
        - updated_at
      title: ClassResponseDTO
      description: DTO for representing a class in API responses.
    ClassUpdateDTO:
      properties:
        class_name:
          anyOf:
            - type: string
              maxLength: 100
              minLength: 1
            - type: 'null'
          title: Class Name
        grade_level:
          anyOf:
            - type: string
              maxLength: 50
            - type: 'null'
          title: Grade Level
      type: object
      title: ClassUpdateDTO
      description: DTO for updating an existing class. All fields are optional.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_ClassResponseDTO_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ClassResponseDTO'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
      type: object
      required:
        - items
        - total
        - page
        - size
      title: PaginatedResponse[ClassResponseDTO]
...
    UserResponseDTO:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        email:
          type: string
          format: email
          title: Email
        first_name:
          anyOf:
            - type: string
            - type: 'null'
          title: First Name
        last_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Last Name
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
        - user_id
        - email
        - role
      title: UserResponseDTO
      description: |-
        A DTO for representing basic user details, often nested in other responses
        (e.g., details of students in a class list).
        It omits sensitive information like password_hash.
    UserRole:
      type: string
      enum:
        - student
        - teacher
        - parent
        - admin
      title: UserRole
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login